Chore(backend): Rename License Cleaner package
