+++ src/main/java/com/philips/research/bombase/core/license_cleaner/domain/package-info.java
@@ -4,4 +4,4 @@
  */
 
 @pl.tlinkowski.annotation.basic.NonNullPackage
-package com.philips.research.bombase.core.license.domain;
+package com.philips.research.bombase.core.license_cleaner.domain;
