+++ src/main/java/com/philips/research/bombase/core/meta/registry/MetaRegistry.java
@@ -10,7 +10,6 @@ import com.philips.research.bombase.core.meta.MetaStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import pl.tlinkowski.annotation.basic.NullOr;
 
 import java.util.HashSet;
 import java.util.Map;
@@ -80,14 +79,12 @@ public class MetaRegistry {
         }
     }
 
-    private void notifyListeners(PackageURL purl, Set<Field> modifiedFields, Map<Field, @NullOr Object> values) {
-        listeners.forEach(l -> {
-            l.onUpdated(purl, modifiedFields, values)
-                    .ifPresent(task -> {
-                        LOG.info("Scheduled {} task for {}", nameFor(l), purl);
-                        runner.execute(purl, task, this::cascadeListeners);
-                    });
-        });
+    private void notifyListeners(PackageURL purl, Set<Field> modifiedFields, Map<Field, Object> values) {
+        listeners.forEach(l -> l.onUpdated(purl, modifiedFields, values)
+                .ifPresent(task -> {
+                    LOG.info("Scheduled {} task for {}", nameFor(l), purl);
+                    runner.execute(purl, task, this::cascadeListeners);
+                }));
     }
 
     /**
@@ -102,6 +99,6 @@ public class MetaRegistry {
          * @param values  current package metadata
          * @return (optional) operation to queue for execution
          */
-        Optional<Consumer<PackageAttributeEditor>> onUpdated(PackageURL purl, Set<Field> updated, Map<Field, @NullOr Object> values);
+        Optional<Consumer<PackageAttributeEditor>> onUpdated(PackageURL purl, Set<Field> updated, Map<Field, Object> values);
     }
 }
